% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{cfit_theme}
\alias{cfit_theme}
\title{compfit ggplot2 theme}
\usage{
cfit_theme(base_size = 11)
}
\arguments{
\item{base_size}{Base font size.}
}
\value{
A ggplot2 theme object.
}
\description{
A minimal ggplot2 theme used across compfit plots.
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(wt, mpg)) + geom_point() + cfit_theme()
}
