% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{compfit-accessors}
\alias{compfit-accessors}
\alias{get_loss}
\alias{get_model}
\alias{get_data}
\alias{get_time_grid}
\alias{get_bounds}
\alias{get_solver}
\alias{get_compartmental_function}
\alias{get_julia_code}
\alias{get_expressions}
\alias{get_states_and_params}
\alias{get_point}
\alias{get_samples}
\title{Accessors for a compartmental fit}
\usage{
get_loss(fit)

get_model(fit)

get_data(fit)

get_time_grid(fit)

get_bounds(fit)

get_solver(fit)

get_compartmental_function(fit)

get_julia_code(fit)

get_expressions(fit)

get_states_and_params(fit)

get_point(fit)

get_samples(fit)
}
\arguments{
\item{fit}{A `"compartmentalFit"` object.}
}
\value{
The requested component of the fit.
}
\description{
Extract components of a `"compartmentalFit"` object.
}
\examples{
\dontrun{
# fit from fitCompartmentalModel()
get_point(fit)
get_julia_code(fit)
}
}
