% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesPost.R
\name{posterior_means}
\alias{posterior_means}
\title{Posterior means (natural scale)}
\usage{
posterior_means(fit)
}
\arguments{
\item{fit}{A Bayesian `"compartmentalFit"` object.}
}
\value{
A named numeric vector of posterior means.
}
\description{
Posterior means (natural scale)
}
\examples{
\dontrun{
posterior_means(fit_b)   # fit_b a method = "bayes" fit
}
}
