% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{summary.compartmentalFit}
\alias{summary.compartmentalFit}
\title{Summary of a compartmental fit}
\usage{
\method{summary}{compartmentalFit}(object, ...)
}
\arguments{
\item{object}{A `"compartmentalFit"` object.}

\item{...}{Unused.}
}
\value{
`object`, invisibly (called for the printed summary).
}
\description{
Summary of a compartmental fit
}
\examples{
\dontrun{
summary(fit)   # fit from fitCompartmentalModel()
}
}
